Serialization to and from C structs.
