+++ sil.cabal
@@ -49,9 +49,23 @@ test-suite sil-test
   type:                exitcode-stdio-1.0
   hs-source-dirs:      test
   main-is:             Spec.hs
+  other-modules:       Common
   build-depends:       base
                      , sil
                      , strict
+                     , QuickCheck
+  ghc-options:         -threaded -rtsopts -with-rtsopts=-N
+  default-language:    Haskell2010
+
+test-suite sil-serializer-test
+  type:                exitcode-stdio-1.0
+  hs-source-dirs:      test/
+  main-is:             SerializerSpec.hs
+  other-modules:       Common
+  build-depends:       base
+                     , sil
+                     , strict
+                     , hspec
                      , QuickCheck
   ghc-options:         -threaded -rtsopts -with-rtsopts=-N
   default-language:    Haskell2010
