+++ src/SIL/Serializer.hs
@@ -51,10 +51,47 @@ data CPLeft   = CPLeft   CTypeId (Ptr CRep) deriving(Show, Generic, GStorable)
 data CPRight  = CPRight  CTypeId (Ptr CRep) deriving(Show, Generic, GStorable) 
 data CTrace   = CTrace   CTypeId (Ptr CRep) deriving(Show, Generic, GStorable) 
 
--- | Obtains (hopefully valid) IExpr from C representation.
---fromC :: Ptr CRoot -> IO IExpr
---fromC ptr = do
 
+-- | Obtains (hopefully valid) IExpr from C representation.
+fromC :: Ptr CRoot -> IO IExpr
+fromC ptr = do
+    (CRoot t v) <- peek ptr
+    fromC' t v
+    
+fromC' :: CTypeId -> Ptr CRep -> IO IExpr
+fromC' type_id ptr = case type_id of
+    0    -> return Zero
+    1    -> do
+        (CPair l_type r_type l_val r_val) <- peek $ castPtr ptr
+        Pair <$> fromC' l_type l_val <*> fromC' r_type r_val
+    2    -> return Env
+    3    -> do
+        (CSetEnv t v) <- peek $ castPtr ptr
+        SetEnv <$> fromC' t v
+    4    -> do
+        (CDefer t v) <- peek $ castPtr ptr
+        Defer <$> fromC' t v
+    5    -> do
+        (CTwiddle t v) <- peek $ castPtr ptr
+        Twiddle <$> fromC' t v
+    6    -> do
+        (CAbort t v) <- peek $ castPtr ptr
+        Abort <$> fromC' t v
+    7    -> do
+        (CGate t v) <- peek $ castPtr ptr
+        Gate <$> fromC' t v
+    8    -> do
+        (CPLeft t v) <- peek $ castPtr ptr
+        PLeft <$> fromC' t v
+    9    -> do
+        (CPRight t v) <- peek $ castPtr ptr
+        PRight <$> fromC' t v
+    10   -> do
+        (CTrace t v) <- peek $ castPtr ptr
+        Trace <$> fromC' t v
+    otherwise -> error "SIL.Serializer.fromC': Invalid type id - possibly corrupted data."
+    
+    
 
 -- | Saves the IExpr as a C representation.
 toC :: IExpr -> IO (Ptr CRoot)
@@ -85,7 +122,7 @@ toC' (Pair e1 e2) ptr_type ptr_value = do
     poke ptr_value $ castPtr value
     toC' e1 ptr_left_type ptr_left_value
     toC' e2 ptr_right_type ptr_right_value
-toC' (Env)  ptr_type ptr_value = poke ptr_type zero_type >> poke ptr_value nullPtr
+toC' (Env)  ptr_type ptr_value = poke ptr_type env_type >> poke ptr_value nullPtr
 toC' (SetEnv e) ptr_type ptr_value = do
     value <- malloc :: IO (Ptr CSetEnv)
     let align      = alignment (undefined :: CSetEnv)
@@ -99,7 +136,7 @@ toC' (Defer e) ptr_type ptr_value = do
     let align      = alignment (undefined :: CDefer)
         next_type  = castPtr value
         next_value = castPtr $ value `plusPtr` align
-    poke ptr_type setenv_type
+    poke ptr_type defer_type
     poke ptr_value $ castPtr value
     toC' e next_type next_value
 toC' (Twiddle e) ptr_type ptr_value = do
@@ -107,7 +144,7 @@ toC' (Twiddle e) ptr_type ptr_value = do
     let align      = alignment (undefined :: CTwiddle)
         next_type  = castPtr value
         next_value = castPtr $ value `plusPtr` align
-    poke ptr_type setenv_type
+    poke ptr_type twiddle_type
     poke ptr_value $ castPtr value
     toC' e next_type next_value
 toC' (Abort e) ptr_type ptr_value = do
@@ -115,7 +152,7 @@ toC' (Abort e) ptr_type ptr_value = do
     let align      = alignment (undefined :: CAbort)
         next_type  = castPtr value
         next_value = castPtr $ value `plusPtr` align
-    poke ptr_type setenv_type
+    poke ptr_type abort_type
     poke ptr_value $ castPtr value
     toC' e next_type next_value
 toC' (Gate e) ptr_type ptr_value = do
@@ -123,7 +160,7 @@ toC' (Gate e) ptr_type ptr_value = do
     let align      = alignment (undefined :: CGate)
         next_type  = castPtr value
         next_value = castPtr $ value `plusPtr` align
-    poke ptr_type setenv_type
+    poke ptr_type gate_type
     poke ptr_value $ castPtr value
     toC' e next_type next_value
 toC' (PLeft e) ptr_type ptr_value = do
@@ -131,7 +168,7 @@ toC' (PLeft e) ptr_type ptr_value = do
     let align      = alignment (undefined :: CPLeft)
         next_type  = castPtr value
         next_value = castPtr $ value `plusPtr` align
-    poke ptr_type setenv_type
+    poke ptr_type pleft_type
     poke ptr_value $ castPtr value
     toC' e next_type next_value
 toC' (PRight e) ptr_type ptr_value = do
@@ -139,7 +176,7 @@ toC' (PRight e) ptr_type ptr_value = do
     let align      = alignment (undefined :: CPRight)
         next_type  = castPtr value
         next_value = castPtr $ value `plusPtr` align
-    poke ptr_type setenv_type
+    poke ptr_type pright_type
     poke ptr_value $ castPtr value
     toC' e next_type next_value
 toC' (Trace e) ptr_type ptr_value = do
@@ -147,6 +184,6 @@ toC' (Trace e) ptr_type ptr_value = do
     let align      = alignment (undefined :: CTrace)
         next_type  = castPtr value
         next_value = castPtr $ value `plusPtr` align
-    poke ptr_type setenv_type
+    poke ptr_type trace_type
     poke ptr_value $ castPtr value
     toC' e next_type next_value
